Move stackstorm_api.js to src/lib/
