+++ src/lib/stackstorm_api.js
@@ -33,10 +33,10 @@ var _ = require('lodash'),
   util = require('util'),
   env = _.clone(process.env),
   Promise = require('rsvp').Promise,
-  utils = require('./lib/utils.js'),
-  formatCommand = require('./lib/format_command.js'),
-  CommandFactory = require('./lib/command_factory.js'),
-  adapters = require('./lib/adapters'),
+  utils = require('./utils.js'),
+  formatCommand = require('./format_command.js'),
+  CommandFactory = require('./command_factory.js'),
+  adapters = require('./adapters'),
   st2client = require('st2client'),
   uuid = require('uuid')
   ;
